fix: move Symbol and SymbolKind to submodule
